Trust imported keys when verifying signatures
This is required when signature verification is enabled as git will
otherwise give an 'U' status back for the signature (good signature
with unknown validity) instead of the 'G' we are aiming for (good
(valid) signature).

We apply this blindly on all keys known to gpg as all keys are added
by the user and thus already trusted.